Feat: add basic functionality to log in out
Installed faker gem. It's used to generate random data, like names,
emails, colors, etc. Seeded the db with some functions that utilize
this gem. For the password field, set all users with the same password.
This is done because when setting a random password, I couldn't obtain
this password to log in. Obiously in production each user will provide
its own password.

Added validations for the User columns. Also created a users
controllers. Created an index action for the moment. This action is the
root and will display a list of all users registered thanks to the view.

Also added a require_authenticable method for all pages. And when a user
has signed in, it will see a link to log out. The log out is placed on
a partial and its rendered in layouts/application.html.erb