Feat: let users comment posts
Created and migrated Comment model. Only has content, user_id and
post_id columns.

Set the associations between the models: A user has many comments, a
post has many comments and a comment belongs to both user and comments.

Also created a comments controller with just the new and create actions.
Set the routes file.

In the posts show view, added a form to let the user post a comment.
This form contains a text_area to enter a comment and two hidden fields
to send the user_id and post_id to correctly create the comment record.

Changed the render statements in likes controller to redirect_to with a
different status than the default one.