Feat: add friends requests, navbar
Created and migrated a Relationship model now. Also set associations
between This model and the User. The Relationship model helps register
the users that send a friend request to another. It only permits one
record per pair of users. This is done thanks to validations, an index
and a method that checks inverse combinations between two user ids.

There is a relationships controller with actions to create the friend
requests.

There's also a button inside a form to send the friend request. This
code is stored in a partial to be used both in the show page of each
user and below the user name in the users page.

Another important thing about the friend request button, is that it's
hidden to the current user and to another user that he has sent or
received a friend request earlier.

Changed the root page to the Timeline page.

There is now an incomplete navbar that contains some links to other
sites like the users page, the timeline, the users profile and a sign
out link. It needs to have a notifications button to show friend
requests received. This navbar is inside another partial, that's
rendered on the application view, to be viewed in all the pages.

The show page now contains a lists of links to all the posts that the
current user has created.