Feat: add buttons to handle friend requests
There are new actions for the relationships controllers: index, update
and destroy. Added to routes file.

In the new index view for relationships, it's the notifications feature
but in another page instead of the same page. It lists all the pending
friend requests that a user has received, along with two buttons, each
one with a form to confirm or delete the friend request.

There's also a notifications link in the user navbar.