Refactor: Improve timeline feature
The timeline (posts index view) now contains post from all the friends
besides the current user's posts.

There are a few class methods in Relationship model to fetch those
entries where the relationship type is friends and the user id pass
matches the requester or requested id. Then another methods retrieve
the other user id. And finally, a method to gather the results of
previous methods.

In the Post Model, there is a class method to obtain all the posts
based on an array of user_ids received (e.g. the friends).

In the posts controller, there is a friends instance variable that calls
the Relationship#friends and appends the current user id. The posts
instance variable uses the new method defined on the model.

In the index view, now it displays the post author, the date of the post
creation, links on the post author name and post title and a count of
comments for each post.